Moved error detection code to its own module; fixed imports; some minor tech debt repair
