+++ kgx/graph_operations/summarize_graph.py
@@ -9,7 +9,7 @@ import yaml
 from json import dump
 from json.encoder import JSONEncoder
 
-from kgx import ErrorDetecting, ErrorType, MessageLevel
+from kgx.error_detection import ErrorType, MessageLevel, ErrorDetecting
 from kgx.utils.kgx_utils import GraphEntityType
 from kgx.graph.base_graph import BaseGraph
 from kgx.prefix_manager import PrefixManager
